refactor(health): migrate CRUD to Pydantic schemas and improve processing robustness (mypy) (#734)
## What does this PR do?

Continues the #678 refactor series by standardizing the `health` module.

## Related issue

#678

## Type of change

- [ ] Bug fix
- [ ] New feature (exception during freeze; link prior discussion/acknowledgment)
- [ ] Documentation
- [x] Refactor / code quality
- [ ] Translation
- [ ] Other (describe below)

## Checklist

- [x] This PR is focused on a single concern
- [x] An issue was opened and acknowledged before writing this code (if not a small fix)
- [ ] If this PR introduces a new feature during freeze, it was discussed and explicitly acknowledged first
- [x] The PR is under ~300 lines changed, excluding lock files, migrations, and generated code
- [ ] Tests have been added or updated where applicable
- [ ] Documentation has been updated where applicable
- [x] Commit messages follow Conventional Commits format

## Notes for the reviewer

All changes follow the same pattern established in previous #678 PRs (e.g. `server-settings`, `notifications`, `users`, etc). The migration processing guards are defensive improvements.

Reviewed-on: https://codeberg.org/endurain-project/endurain/pulls/734

+++ backend/app/health/health_sleep/sleep_scoring.py
@@ -356,11 +356,11 @@ def calculate_sleep_stress_score(
     if stress_value <= 25:  # Rest state
         base_score = 100
     elif stress_value <= 50:  # Low stress
-        base_score = 90 - ((stress_value - 25) / 25 * 20)
+        base_score = 90 - int((stress_value - 25) / 25 * 20)
     elif stress_value <= 75:  # Medium stress
-        base_score = 70 - ((stress_value - 50) / 25 * 20)
+        base_score = 70 - int((stress_value - 50) / 25 * 20)
     else:  # High stress
-        base_score = 50 - ((stress_value - 75) / 25 * 50)
+        base_score = 50 - int((stress_value - 75) / 25 * 50)
 
     # Penalty for restless moments
     restless_penalty = 0
